Modify Delete User

+++ app/Http/Controllers/UserController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\DataDiri;
 use App\Models\ModelRequestRole;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Route;
+use Illuminate\Support\Facades\Storage;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
@@ -87,9 +89,13 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
+        $dataDiri = DataDiri::where("user_id", $user->id)->first();
+        if (Storage::disk('public')->delete($dataDiri->profile_picture)) {
+            $user->delete();
+            return redirect()->route('users.index')->with('success', 'User Deleted successfully.');
+        }
 
-        return redirect()->route('users.index')->with('success', 'Event Deleted successfully.');
+        return redirect()->route('users.index')->with('error', 'User Failed Deleted successfully.');
     }
 
     public function pendingSubmission()
